app/requirements.docker.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66

timm>=1.0.7

pillow>=10.0
opencv-python-headless>=4.8
albumentations>=1.4.7

scikit-learn>=1.3